refactor: move vue-plugin to services
