+++ packages/ui/src/components/va-modal/plugin/index.ts
@@ -1,5 +1,5 @@
 import { App } from 'vue'
-import { defineVuesticPlugin, defineGlobalProperty } from '../../../vuestic-plugin/utils'
+import { defineVuesticPlugin, defineGlobalProperty } from '../../../services/vue-plugin/utils'
 import { createModalInstance } from '../modal'
 import { ModalOptions } from '../types'
