+++ packages/ui/src/services/color-config/plugin/index.ts
@@ -1,5 +1,5 @@
 import { createColorConfigPlugin } from './create-color-config-plugin'
-import { defineGlobalProperty, defineVuesticPlugin } from '../../../vuestic-plugin/utils'
+import { defineGlobalProperty, defineVuesticPlugin } from '../../vue-plugin/utils'
 
 /** Creates color css variables and reactively updates on ColorConfig changes. */
 export const ColorConfigPlugin = defineVuesticPlugin(() => ({
