+++ packages/ui/src/services/vue-plugin/composables.ts
@@ -1,0 +1,1 @@
+export { useElementTextColor } from '../../composables/useElementTextColor'
