+++ packages/ui/src/services/vue-plugin/create-vuestic/create-vuestic.ts
@@ -1,9 +1,9 @@
-import type { PartialGlobalConfig } from '../../services/global-config/types'
+import type { PartialGlobalConfig } from '../../global-config/types'
 import { defineVuesticPlugin, usePlugin } from '../utils'
-import { GlobalConfigPlugin, VaDropdownPlugin, VaToastPlugin, VaModalPlugin, ColorConfigPlugin, BreakpointConfigPlugin, CachePlugin } from '../vuestic-plugins'
-import * as vuesticComponents from '../vuestic-components'
-import type { VuesticComponents } from '../global-components'
-import { getCurrentApp, setCurrentApp } from '../../services/current-app'
+import { GlobalConfigPlugin, VaDropdownPlugin, VaToastPlugin, VaModalPlugin, ColorConfigPlugin, BreakpointConfigPlugin, CachePlugin } from '../plugins'
+import * as vuesticComponents from '../components'
+import type { VuesticComponents } from '../types/components'
+import { setCurrentApp } from '../../current-app'
 
 // Declare all components globally
 declare module 'vue' {
@@ -29,9 +29,9 @@ export const createVuestic = defineVuesticPlugin((options: { config?: PartialGlo
     // These plugins have dependant plugins, so have to be registered first.
     usePlugin(app, GlobalConfigPlugin(config))
     usePlugin(app, CachePlugin)
+    usePlugin(app, ColorConfigPlugin)
 
     usePlugin(app, BreakpointConfigPlugin)
-    usePlugin(app, ColorConfigPlugin)
     usePlugin(app, VaDropdownPlugin)
     usePlugin(app, VaToastPlugin)
     usePlugin(app, VaModalPlugin)
