+++ packages/ui/src/services/vue-plugin/create-vuestic/index.ts
