+++ packages/ui/src/services/vue-plugin/index.ts
@@ -1,0 +1,4 @@
+export * from './plugins'
+export * from './create-vuestic'
+export * from './types/components'
+export * from './composables'
