+++ packages/ui/src/services/vue-plugin/plugins.ts
@@ -1,0 +1,7 @@
+export { GlobalConfigPlugin } from '../global-config/plugin'
+export { ColorConfigPlugin } from '../color-config/plugin'
+export { BreakpointConfigPlugin } from '../breakpoint/plugin'
+export { VaToastPlugin } from '../../components/va-toast/plugin'
+export { VaDropdownPlugin } from '../../components/va-dropdown/plugin'
+export { VaModalPlugin } from '../../components/va-modal/plugin'
+export { CachePlugin } from '../cache/plugin/index'
