+++ packages/ui/src/services/vue-plugin/types/index.ts
@@ -1,0 +1,14 @@
+import type { App, Plugin, Component } from 'vue'
+
+export type VuesticComponent = Component
+
+type PluginInstallFn<ARGS = []> = (app: App, ...options: ARGS[]) => void
+
+/** Vuestic internal plugin */
+export type VuesticPlugin = {
+  install: PluginInstallFn<[]>,
+} & Plugin
+
+export type VuesticPluginFabric<O extends any[] = []> = (...args: O) => VuesticPlugin
+
+export * from './components'
