+++ packages/ui/src/services/vue-plugin/utils/define-vuestic-plugin.ts
@@ -1,0 +1,3 @@
+import { VuesticPluginFabric } from '../types'
+
+export const defineVuesticPlugin = <O extends any[]>(fabric: VuesticPluginFabric<O>) => fabric
