+++ packages/ui/src/services/vue-plugin/utils/global-properties.ts
