--- packages/ui/src/vuestic-plugin/index.ts
@@ -1,4 +1,0 @@
-export * from './vuestic-plugins'
-export * from './create-vuestic'
-export * from './global-components'
-export * from './vuestic-composables'
