--- packages/ui/src/vuestic-plugin/utils/define-vuestic-plugin.ts
@@ -1,1 +1,0 @@
-export { defineVuesticPlugin } from '../types'
