--- packages/ui/src/vuestic-plugin/vuestic-composables.ts
@@ -1,1 +1,0 @@
-export { useElementTextColor } from '../composables/useElementTextColor'
