--- packages/ui/src/vuestic-plugin/vuestic-plugins.ts
@@ -1,7 +1,0 @@
-export { GlobalConfigPlugin } from '../services/global-config/plugin'
-export { ColorConfigPlugin } from '../services/color-config/plugin'
-export { BreakpointConfigPlugin } from '../services/breakpoint/plugin'
-export { VaToastPlugin } from '../components/va-toast/plugin'
-export { VaDropdownPlugin } from '../components/va-dropdown/plugin'
-export { VaModalPlugin } from '../components/va-modal/plugin'
-export { CachePlugin } from '../services/cache/plugin/index'
